+++ packages/graphics-engine/src/index.js
@@ -9,3 +9,5 @@ export {
 export { DocumentHistory } from "./history.js";
 
 export { getByPath, resolveTemplate, resolveElement, resolveDocument, listBindingTokens } from "./data-binding.js";
+
+export { listBroadcastComponents, createBroadcastComponent, createScriptureComponent, createLowerThirdComponent } from "./broadcast-components.js";